+++ src/test/java/com/github/davidmc24/gradle/plugin/avro/V2TaskTest.java
@@ -1,17 +1,15 @@
 package com.github.davidmc24.gradle.plugin.avro;
 
-import com.google.common.io.MoreFiles;
 import com.google.common.io.Resources;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.List;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.List;
 
 public class V2TaskTest {
     @TempDir
@@ -20,28 +18,28 @@ public class V2TaskTest {
     @Test
     void supportsVersionArgument() throws Exception {
         String buildFileContent = """
-        plugins { id "com.github.davidmc24.gradle.plugin.avro-base" }
-        repositories {
-            mavenCentral()
-            maven { url = "https://oss.sonatype.org/content/repositories/snapshots/" }
-        }
-        
-        configurations {
-            avroSupplement
-        }
-        
-        dependencies {
-            avroSupplement "com.github.davidmc24.avro.supplement:1.10.1:0.0.1-SNAPSHOT"
-            avroSupplement "org.slf4j:slf4j-simple:1.7.30"
-        }
-        
-        tasks.register("v2", com.github.davidmc24.gradle.plugin.avro.V2Task) {
-            classpath.from(configurations.avroSupplement)
-            source("src/main/avro")
-            include("**/*.avsc")
-            destinationDirectory.set(file("build/generated-main-avro-java"))
-        }
-        """.stripIndent();
+            plugins { id "com.github.davidmc24.gradle.plugin.avro-base" }
+            repositories {
+                mavenCentral()
+                maven { url = "https://oss.sonatype.org/content/repositories/snapshots/" }
+            }
+            
+            configurations {
+                avroSupplement
+            }
+            
+            dependencies {
+                avroSupplement "com.github.davidmc24.avro.supplement:1.10.1:0.0.1-SNAPSHOT"
+                avroSupplement "org.slf4j:slf4j-simple:1.7.30"
+            }
+            
+            tasks.register("v2", com.github.davidmc24.gradle.plugin.avro.V2Task) {
+                classpath.from(configurations.avroSupplement)
+                source("src/main/avro")
+                include("**/*.avsc")
+                destinationDirectory.set(file("build/generated-main-avro-java"))
+            }
+            """.stripIndent();
         String expectedJava = "moo"; // TODO: extract
 
         Path testProjectDir = tempDir.resolve("project");
@@ -55,7 +53,6 @@ public class V2TaskTest {
         Files.write(avroFile, Resources.toByteArray(Resources.getResource("examples/inline/Cat.avsc")));
 
         String gradleVersion = "6.8.2"; // TODO: parameterize
-//        List<String> args = List.of("--info", "tasks", "--all");
         List<String> args = List.of("--info", "v2");
         GradleRunner runner = GradleRunner.create()
             .withProjectDir(testProjectDir.toFile())